+++ packages/dart/test/crypto_test.dart
@@ -34,13 +34,13 @@ void main() {
       expect(decodeBase64('MTIzNDU2Nzg5MFRlc3Q='), '1234567890Test');
     });
 
-    test('strToNumberHash', () {
-      expect(strToNumberHash(''), 0);
-      expect(strToNumberHash(' '), 32);
-      expect(strToNumberHash('abc'), 96354);
-      expect(strToNumberHash('Hello'), 69609650);
-      expect(strToNumberHash('hello'), 99162322);
-      expect(strToNumberHash('ABCDEFGHIJKLMNOPQRSTUVWXYZ' * 10000), 285059024);
+    test('numberHash', () {
+      expect(numberHash(''), 0);
+      expect(numberHash(' '), 32);
+      expect(numberHash('abc'), 96354);
+      expect(numberHash('Hello'), 69609650);
+      expect(numberHash('hello'), 99162322);
+      expect(numberHash('ABCDEFGHIJKLMNOPQRSTUVWXYZ' * 10000), 285059024);
     });
   });
 }