+++ packages/javascript/lib/node/crypto/sha256Hash.ts
@@ -1,5 +1,8 @@
 import { createHash } from 'crypto';
+import type { BinaryToTextEncoding } from 'node:crypto';
 
-export function sha256Hash(str: string): string {
-	return createHash('sha256').update(str).digest('hex');
+export function sha256Hash(str: string, encoding?: BinaryToTextEncoding): string {
+	return createHash('sha256')
+		.update(str)
+		.digest(encoding ?? 'hex');
 }